Separate graphics from logic
