+++ .travis.yml
@@ -11,12 +11,13 @@ addons:
 
 install:
   - pip install --upgrade pip
-  - pip install -r requirements.txt
   
-  # PySide installation
+  # Preferred PySide installation, it is faster
   # Based on http://stackoverflow.com/a/24545890/478949
   - pip install PySide --no-index --find-links https://parkin.github.io/python-wheelhouse/;
   - python ~/virtualenv/python${TRAVIS_PYTHON_VERSION}/bin/pyside_postinstall.py -install
+  
+  - pip install -r requirements.txt
 
 script:
   - python run_tests.py
