+++ pyblish_lite/util.py
@@ -1,5 +1,7 @@
 import os
 
+from Qt import QtCore
+
 root = os.path.dirname(__file__)
 
 
@@ -17,3 +19,23 @@ def get_asset(*path):
     """
 
     return os.path.join(root, *path)
+
+
+def defer(delay, func):
+    """Append artificial delay to `func`
+
+    This aids in keeping the GUI responsive, but complicates logic
+    when producing tests. To combat this, the environment variable ensures
+    that every operation is synchonous.
+
+    Arguments:
+        delay (float): Delay multiplier; default 1, 0 means no delay
+        func (callable): Any callable
+
+    """
+
+    delay *= float(os.getenv("PYBLISH_DELAY", 1))
+    if delay > 0:
+        return QtCore.QTimer.singleShot(delay, func)
+    else:
+        return func()
