+++ tests/__init__.py
@@ -2,12 +2,12 @@ import os
 import sys
 
 import pyblish_lite
-from Qt import QtWidgets
+from Qt import QtCore
 
 # Remove artificial delay from GUI
 os.environ["PYBLISH_DELAY"] = "0"
 
 self = sys.modules[__name__]
-self.app = QtWidgets.QApplication.instance()
-self.app = self.app or QtWidgets.QApplication(sys.argv)
+self.app = QtCore.QCoreApplication.instance()
+self.app = self.app or QtCore.QCoreApplication(sys.argv)
 self.module = pyblish_lite
